fix: test workflow, test files, package updates
