+++ .github/workflows/test.yml
@@ -1,6 +1,7 @@
 name: Test
 
 on:
+  pull_request:
   push:
     branches:
       - '**'
