+++ package.json
@@ -44,18 +44,18 @@
     ]
   },
   "dependencies": {
-    "@sentry/node": "^6.13.2",
-    "@sentry/tracing": "^6.13.2",
+    "@sentry/node": "^6.13.3",
+    "@sentry/tracing": "^6.13.3",
     "fastify-plugin": "^3.0.0"
   },
   "devDependencies": {
-    "@commitlint/cli": "^13.1.0",
-    "@commitlint/config-conventional": "^13.1.0",
+    "@commitlint/cli": "^13.2.0",
+    "@commitlint/config-conventional": "^13.2.0",
     "eslint": "^7.32.0",
     "eslint-plugin-markdown": "^2.2.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "fastify": "^3.21.3",
+    "fastify": "^3.22.0",
     "fastify-jwt": "^3.1.0",
     "husky": "^7.0.2",
     "prettier": "^2.4.1",
